+++ corbeans-corda-webserver/src/test/kotlin/com/github/manosbatsis/corbeans/corda/webserver/ActuatorIntegrationTest.kt
@@ -1,0 +1,85 @@
+/*
+ *     Corbeans: Corda integration for Spring Boot
+ *     Copyright (C) 2018 Manos Batsis
+ *
+ *     This library is free software; you can redistribute it and/or
+ *     modify it under the terms of the GNU Lesser General Public
+ *     License as published by the Free Software Foundation; either
+ *     version 3 of the License, or (at your option) any later version.
+ *
+ *     This library is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *     Lesser General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Lesser General Public
+ *     License along with this library; if not, write to the Free Software
+ *     Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ *     USA
+ */
+package com.github.manosbatsis.corbeans.corda.webserver
+
+import com.github.manosbatsis.corbeans.test.integration.CorbeansSpringExtension
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNotNull
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
+import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.boot.test.web.client.TestRestTemplate
+import org.springframework.http.HttpStatus
+
+
+/**
+ * Corbeans actuator tests
+ */
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+// Note we are using CorbeansSpringExtension Instead of SpringExtension
+@ExtendWith(CorbeansSpringExtension::class)
+class ActuatorIntegrationTest {
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(ActuatorIntegrationTest::class.java)
+
+    }
+
+    @Autowired
+    lateinit var restTemplate: TestRestTemplate
+
+    @Test
+    fun testInfoContributor() {
+        logger.debug("testInfoContributor, called")
+        val entity = this.restTemplate
+                .getForEntity("/actuator/info", Map::class.java)
+        // Ensure a 200 OK
+        assertEquals(HttpStatus.OK, entity.getStatusCode())
+
+        val body = entity.getBody()
+        assertNotNull(body, "Actuator info must not be null")
+        val corda = body!!["corda"] as Map<*, *>?
+        // Validate corda information
+        validateCordaInfo(corda)
+    }
+
+    @Test
+    fun testCordaEndpoint() {
+        logger.debug("testCordaEndpoint, called")
+        val entity = this.restTemplate
+                .getForEntity("/actuator/corda", Map::class.java)
+        // Ensure a 200 OK
+        assertEquals(HttpStatus.OK, entity.getStatusCode())
+        val corda = entity.getBody()
+        // Validate corda information
+        validateCordaInfo(corda)
+    }
+
+
+    private fun validateCordaInfo(corda: Map<*, *>?) {
+        assertNotNull(corda, "Actuator corda info must not be null")
+        val cordaNodes = corda!!["nodes"] as Map<String, Any>
+        assertNotNull(cordaNodes["partyA"])
+        assertNotNull(cordaNodes["partyB"])
+    }
+
+}